backend/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
django-filter==23.5
drf-yasg==1.21.7
python-decouple==3.8 